4.23

+++ tests/blog_api.test.js
@@ -1,10 +1,72 @@
-const { test, after, describe } = require('node:test')
+const { test, after, describe, beforeEach } = require('node:test')
 const assert = require('node:assert')
 const mongoose = require('mongoose')
 const supertest = require('supertest')
+const jwt = require('jsonwebtoken')
 const app = require('../app')
+const User = require('../models/user')
+const Blog = require('../models/blog')
+const bcrypt = require('bcrypt')
+
 
 const api = supertest(app)
+
+let token = ''
+let testUserId = ''
+
+beforeEach(async () => {
+  // Clear the DB
+  await User.deleteMany({})
+  await Blog.deleteMany({})
+
+  // Create a test user
+  const saltRounds = 10
+  const passwordHash = await bcrypt.hash('password', saltRounds)
+
+  const testUser = new User({
+    username: 'testuser',
+    name: 'Test User',
+    passwordHash: passwordHash,
+  })
+
+  const savedUser = await testUser.save()
+  testUserId = savedUser._id
+
+  // Generate a token for the test user
+  const userForToken = {
+    username: savedUser.username,
+    id: testUserId,
+  }
+
+  token = jwt.sign(userForToken, process.env.SECRET)
+
+  // console.log('Test user created:', savedUser)
+  // console.log('Token:', token)
+
+  // Create to dummy blog posts
+  const blogObjects = [
+    {
+      title: 'First blog post',
+      author: 'John Doe',
+      url: 'http://example.com/first',
+      likes: 10,
+    },
+    {
+      title: 'Second blog post',
+      author: 'Jane Doe',
+      url: 'http://example.com/second',
+      likes: 20,
+    },
+  ]
+  // post the two blogs
+  let blogObject = new Blog(blogObjects[0])
+  blogObject.user = testUserId
+  await blogObject.save()
+  blogObject = new Blog(blogObjects[1])
+  blogObject.user = testUserId
+  await blogObject.save()
+
+})
 
 describe('blogApi', () => {
 
@@ -16,18 +78,9 @@ describe('blogApi', () => {
         .expect('Content-Type', /application\/json/)
     })
 
-    test('there are two blogs', async () => {
+    test('there are two blogs initially', async () => {
       const response = await api.get('/api/blogs')
-
       assert.strictEqual(response.body.length, 2)
-    })
-
-    test('the first blog is about a white whale', async () => {
-      const response = await api.get('/api/blogs')
-
-      const titles = response.body.map(e => e.title)
-      // is the argument truthy
-      assert(titles.includes('Moby Dick'))
     })
 
     test('the unique identifier property of blog posts is named id', async () => {
@@ -57,6 +110,7 @@ describe('blogApi', () => {
       // Post a new blog
       await api
         .post('/api/blogs')
+        .set('Authorization', `Bearer ${token}`)
         .send(newBlog)
         .expect(201)
         .expect('Content-Type', /application\/json/)
@@ -74,7 +128,7 @@ describe('blogApi', () => {
 
       // Delete the added blog post
       const addedBlog = blogsAtEnd.find(blog => blog.title === newBlog.title)
-      await api.delete(`/api/blogs/${addedBlog.id}`).expect(204)
+      await api.delete(`/api/blogs/${addedBlog.id}`).set('Authorization', `Bearer ${token}`).expect(204)
 
       // Verify the number of blogs is back to the initial count
       const finalBlogs = await api.get('/api/blogs')
@@ -92,6 +146,7 @@ describe('blogApi', () => {
       // Post a new blog
       const response = await api
         .post('/api/blogs')
+        .set('Authorization', `Bearer ${token}`)
         .send(newBlog)
         .expect(201)
         .expect('Content-Type', /application\/json/)
@@ -101,7 +156,7 @@ describe('blogApi', () => {
       assert.strictEqual(savedBlog.likes, 0)
 
       // Delete the added blog post
-      await api.delete(`/api/blogs/${savedBlog.id}`).expect(204)
+      await api.delete(`/api/blogs/${savedBlog.id}`).set('Authorization', `Bearer ${token}`).expect(204)
     })
 
     test('a blog without a title property returns status 400 Bad Request', async () => {
@@ -113,6 +168,7 @@ describe('blogApi', () => {
 
       await api
         .post('/api/blogs')
+        .set('Authorization', `Bearer ${token}`)
         .send(newBlog)
         .expect(400)
     })
@@ -126,8 +182,23 @@ describe('blogApi', () => {
 
       await api
         .post('/api/blogs')
+        .set('Authorization', `Bearer ${token}`)
         .send(newBlog)
         .expect(400)
+    })
+
+    test('adding a blog without a token fails with status code 401', async () => {
+      const newBlog = {
+        title: 'Sneaky post',
+        author: 'Who knows',
+        url: 'http://example.com/sneaky',
+        likes: 0
+      }
+
+      await api
+        .post('/api/blogs')
+        .send(newBlog)
+        .expect(401)
     })
   })
 
@@ -143,6 +214,7 @@ describe('blogApi', () => {
       // Add a new blog to ensure there is one to delete
       const addedBlog = await api
         .post('/api/blogs')
+        .set('Authorization', `Bearer ${token}`)
         .send(newBlog)
         .expect(201)
         .expect('Content-Type', /application\/json/)
@@ -150,6 +222,7 @@ describe('blogApi', () => {
       // Delete the newly added blog
       await api
         .delete(`/api/blogs/${addedBlog.body.id}`)
+        .set('Authorization', `Bearer ${token}`)
         .expect(204)
 
       // Verify the blog is actually deleted
@@ -173,6 +246,7 @@ describe('blogApi', () => {
       // Add a new blog to make sure there is something to update
       const addedBlog = await api
         .post('/api/blogs')
+        .set('Authorization', `Bearer ${token}`)
         .send(newBlog)
         .expect(201)
         .expect('Content-Type', /application\/json/)
@@ -187,6 +261,7 @@ describe('blogApi', () => {
       // Update the newly added blog
       const response = await api
         .put(`/api/blogs/${addedBlog.body.id}`)
+        .set('Authorization', `Bearer ${token}`)
         .send(updatedBlog)
         .expect(200)
         .expect('Content-Type', /application\/json/)
@@ -199,7 +274,7 @@ describe('blogApi', () => {
       assert.strictEqual(updatedBlogResponse.likes, updatedBlog.likes)
 
       // Delete the updated blog post
-      await api.delete(`/api/blogs/${updatedBlogResponse.id}`).expect(204)
+      await api.delete(`/api/blogs/${updatedBlogResponse.id}`).set('Authorization', `Bearer ${token}`).expect(204)
     })
   })
 })
